both-mode compiles
